requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
google-generativeai
chromadb
langchain-chroma
langchain_google_genai
marker-pdf
unstructured[md]
llama-parse
fake-useragent
langchain-google-firestore